==> Wildfire/03-resampling.ipynb <==
## Resampling Monthly Data 

import pandas as pd
from IPython.display import clear_output

weather = pd.read_csv('./data/weather_data.csv')
weather['date'] = pd.to_datetime(weather.date)
weather.info()

At this point the weather data is aggregated on a monthly basis.  Since we are predicting the effect that weather might have on forest fires it may be beneficial to look at what the weather has been like in a given county for the months leading up to a potential fire.  If it has not rained in months, for example, this might be a valuable predictive feature.  Therefore we resample some selected features on a 3 month basis and add these new features to our data frame.

county_list = list(weather['county'].unique())

county_dfs = []
for county in county_list:
    clear_output()
    print(county)
    x = weather[weather['county'] == county]
    x.set_index('date', inplace=True)
    county_dfs.append(x)

iters = 0
for df in county_dfs:
    county = df['county'][0]
    trans_df = df[['precip']].rolling(3).sum()
    temp_df = df[['avgtempF', 'humid']].rolling(3).mean()
    temp_df['county'] = county
    temp_df['precip'] = trans_df['precip']
    temp_df.reset_index(inplace = True)
    if iters == 0:
        out = temp_df.copy()
    else:
        out = pd.concat([out, temp_df], ignore_index = True)
    iters += 1
    clear_output()
    print(iters)
    


out['date'] = out['date'].astype(str)
out['date'] = out['date'].map(lambda s: s[0:-3])

weather['date'] = weather['date'].astype(str)
weather['date'] = weather['date'].map(lambda s: s[0:-3])

weather['q_avgtempF'] = out['avgtempF']
weather['q_avghumid'] = out['humid']
weather['q_sumprecip'] = out['precip']


weather.info()

Next we need to merge the fire data to the completed weather data.  Remember from before that the ```UNIT_ID``` feature from the fire data contains some cases where one unit id corresponded to two counties.  To correect for this we assign those fires to only one of the counties so that we can complete the merge.

fire = pd.read_csv('./data/fire_data.csv')

def remove_dash(x):
    if '-' in x:
        county = x.split('- ')
        return county[1]
    else:
        return x
        
fire['UNIT_ID'] = fire['UNIT_ID'].apply(remove_dash)
final = pd.merge(weather, fire, how = 'left', left_on = ['county', 'date'], right_on = ['UNIT_ID', 'ALARM_DATE'])
final.head()

Next we drop columns that will not be used for modeling.

final.drop(columns = ['UNIT_ID', 'ALARM_DATE', 'CONT_DATE', 'SHAPE_Length', 'SHAPE_Area', 'REPORT_AC'], inplace = True)

final.head()

final.info()

There are some missing values that need to be imputed.  Some of these, such as where the ```GIS_ACRES``` is null we can just change that to 0 since the null represents that there was no fire.  For the missing causes we will assign a 0 if there was no fire, but will assign the most common cause for the instances where we do have a fire, but a null in the cause.

final['FIRE_NAME'].fillna('no_fire', inplace = True)

final['GIS_ACRES'].fillna(0, inplace = True)

mode_cause = final['CAUSE'].value_counts().index[0]

x , y = final.shape
for i in range(x):
    clear_output()
    print(f"{i} of {x-1}")
    if final.iloc[i,-1] != 0 and pd.isnull(final.iloc[i,-2]):
        final.iloc[i,-2] = mode_cause

final['CAUSE'].fillna(0, inplace = True)

Since we used a 3 month average for some of the resampled fields there are two records for each county that contain null weather data corresponding to the first two months in our time interval.  These can be dropped.

final.dropna(inplace = True)

final.info()

All of our data was provided in imperial units with the exception of snow, which was given in centimeters.  These units are converted next.

final['totalSnow'] = final['totalSnow_cm'] / 2.54
final.drop(columns = ['totalSnow_cm'], inplace = True)

Finally, the columns of our data frame are reordered to collect all of the weather data in contiguous columns simply for ease of readability.  The dataset is now complete and saved as ```combined.csv``` for use in eda and modeling.

final = final[['date', 'county', 'maxtempF', 'mintempF', 'avgtempF', 'totalSnow', 'humid', 'wind', 'precip', 
               'q_avgtempF', 'q_avghumid', 'q_sumprecip', 'sunHour', 'FIRE_NAME', 'CAUSE', 'lat', 'long', 'GIS_ACRES']]

final.to_csv('./data/combined.csv', index = False)
